ticket #233 redirection issue fixed.
